Release HIntLib-0.0.13 from http://mint.sbg.ac.at/HIntLib/ 0.0.13	March 14, 2008	1.4 MB	HIntLib-0.0.13.tar.gz

+++ src/digitalnetgen.cpp
@@ -30,10 +30,10 @@
  *  Digital Net
  */
 
-HIntLib::DigitalNet::DigitalNet (unsigned base, unsigned _m)
+HIntLib::DigitalNet::DigitalNet (int base, int _m)
    : m (std::min (
         _m,
-        unsigned(logInt (std::numeric_limits<Index>::max(), Index(base))))),
+        logInt (std::numeric_limits<Index>::max(), Index(base)))),
      size (powInt (base, m))
 {}
